third_party/WebKit/LayoutTests/inspector/sass/test-ast-scss-4.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="./sass-test.js"></script>
<style>
pre {
    font-family: monospace;
}
</style>
<script>

function test()
{
    InspectorTest.evaluateInPage("getSASS()", onSASS);

    function onSASS(result)
    {
        InspectorTest.parseSCSS("", result.value)
            .then(InspectorTest.dumpAST)
            .then(InspectorTest.validateASTRanges)
            .catch(console.error.bind(console))
            .then(InspectorTest.completeTest);
    }
}

function getSASS()
{
    return document.querySelector(".snippet").textContent;
}

</script>
</head>

<body onload="runTest()">
<p>
Verifies AST of SCSS with media query.
</p>
<pre class="snippet">
@media (min-device-width: 720px) {
    body {
        width: 720px;
        /* padding-top: 5em; */
        height: 100%;
        margin: 0 auto;
    }
}

</pre>
</body>
</html>
