third_party/WebKit/LayoutTests/inspector/sass/test-ast-scss-4-expected.txt

Verifies AST of SCSS with media query.

@media (min-device-width: 720px) {
    body {
        width: 720px;
        /* padding-top: 5em; */
        height: 100%;
        margin: 0 auto;
    }
}

=== AST === 
rule 0: "body"
    property 0
        name: "width"
            range: {"startLine":2,"startColumn":8,"endLine":2,"endColumn":13}
        value: " 720px"
            range: {"startLine":2,"startColumn":14,"endLine":2,"endColumn":20}
        range: {"startLine":2,"startColumn":8,"endLine":2,"endColumn":21}
        disabled: false
    property 1
        name: "padding-top"
            range: {"startLine":3,"startColumn":11,"endLine":3,"endColumn":22}
        value: " 5em"
            range: {"startLine":3,"startColumn":23,"endLine":3,"endColumn":27}
        range: {"startLine":3,"startColumn":8,"endLine":3,"endColumn":31}
        disabled: true
    property 2
        name: "height"
            range: {"startLine":4,"startColumn":8,"endLine":4,"endColumn":14}
        value: " 100%"
            range: {"startLine":4,"startColumn":15,"endLine":4,"endColumn":20}
        range: {"startLine":4,"startColumn":8,"endLine":4,"endColumn":21}
        disabled: false
    property 3
        name: "margin"
            range: {"startLine":5,"startColumn":8,"endLine":5,"endColumn":14}
        value: " 0 auto"
            range: {"startLine":5,"startColumn":15,"endLine":5,"endColumn":22}
        range: {"startLine":5,"startColumn":8,"endLine":5,"endColumn":23}
        disabled: false
======
Ranges OK.
